Add files via upload
The app is no longer functioning due to compatibility issues between the versions of scikit-learn and pickle. While I was able to successfully host the app on my local machine, deploying it to the cloud proved to be problematic, which has prevented me from making a fully functional prediction model.